add workaround for weird rendering bug
for some reason, the toolbar icons were all getting placed in the upper
right corner of the screen & were therefore invisible & unusable, even
though their parent scroll view was in the right place. this workaround
just skips the ListView and lays them out directly. this will need to
be changed if we ever want the toolbar to have more sophisticated
behavior, e.g. tap an icon, then have the toolbar’s contents change to
allow selection from a horizontally-scrolling list, as seen in some
implementations of autocomplete on mobile.